forwarding: extract HA redundancy-group resolution into ha.rs
Pure code-motion split of forwarding/mod.rs (#5650). Move the HA
redundancy-group resolution enforcement (enforce_ha_resolution*,
cached_flow_decision_valid, finalize_new_flow_ha_resolution), the
owner-RG attribution helpers (owner_rg_for_flow, owner_rg_for_resolution)
and the demoted/activated owner-RG set diff helpers into
forwarding/ha.rs.

Functions are moved verbatim; no logic, signature, or hot-path change.
pub(super) becomes pub(in crate::afxdp) so the mod.rs glob re-export
keeps all call sites resolving.

Validation: cargo build clean (0 errors).
